Airline price stratagies
This file contains the prices of flights to the destinations that our airline will fly to. Each price displayed is based on a return flight departing on 20/11/2017 and returning on 27/11/2017. Each price returned is the cost for one adult passenger. This file will be used to give and ide of the prices to be used in our flight system.